+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure openstack is always detected as CAPI.
+
 ## [3.0.0] - 2022-03-28
 
 ### Added